requirements.txt
anyio==3.7.1
autopep8==2.0.2
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coloredlogs==15.0.1
cssselect2==0.7.0
ddddocr==1.4.7
defusedxml==0.7.1
exceptiongroup==1.1.3
flatbuffers==23.5.26
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
mpmath==1.3.0
numpy==1.25.2
onnxruntime==1.15.1
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==9.5.0
platformdirs==3.10.0
playwright==1.37.0
protobuf==4.24.0
pycodestyle==2.11.0
pycparser==2.21
pyee==9.0.4
pytesseract==0.3.10
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
sympy==1.12
tinycss2==1.2.1
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
webencodings==0.5.1
yapf==0.40.1
zipp==3.16.2
